Update utils.service.ts with new power calculation
Updated lookup tables to reflect in-game power scaling. Created new getCharcacterPower function that reflects how the variables are used to scale character stats, and set initial coefficients to scale power range to roughly that shown in-game. All characters are treated equally, but a parameter is included (dirtyDozenCoeff) that could be used to distinguish inherent power differences between characters.