Fix LRE crash for invalid characters
The LRETile component was prone to crashing when not supplied with expected attributes for a unit.

This was coming up after I migrated the 'Unknown' character to 'Trajann', as some users had the nnow-orphaned 'Unknown' character in their JSON data.

The component now guards against each required attribute, and avoids rendering those aspects. This allows the invalid character to still be displayed, giving the user a means of deleting it.